### 2019-03-24 #1308 @cMadan
@jg-you @count0 @antoineallard--Hi, would you be interested/available to review this project for publication in JOSS? Thank you for considering!
