### 2019-03-27 #1308 @antoineallard
Never done software reviews before either. I may be able to look at the code in a few weeks. One of the coauthors is a friend. Is that an issue?
