### 2019-03-29 #1308 @cMadan
@jg-you @antoineallard, will get back to you soon. I want to try and get someone who's done a software review also on board and then perhaps have one of you as an additional reviewer.

@count0, no problem, thank you for considering!

@hneutr, can you take a look at the JOSS reviewer list (linked in the first comment in this issue), and see if anyone there might be a good fit? It would be preferable to have one person that has done a software review and has a been through the process here at JOSS before.
