### 2019-03-29 #1308 @hneutr
Hey @cMadan — took a look at the list and, prioritizing for web-involved people with experience in Javascript and python, found three that seemed to me reasonably aligned with the project: krischer, wtgee, and Danbo3004. More than a few others would probably work, too, but these three were the ones that caught!
