### 2019-03-31 #1308 @cMadan
@krischer @danbo3004 @wtgee, would you be interested/available to review this project for publication in JOSS? Thank you for considering!
