### 2019-05-11 #1308 @cMadan
@vc1492a @danlurie @raamana, would you be interested/available to review this project for publication in JOSS? Thank you for considering!
