### 2019-05-13 #1308 @raamana
Thanks Chris for the invite. As much as I love to review this (I myself worked towards making such tools - look at [draw3dnx in graynet](https://github.com/raamana/graynet/blob/master/graynet/vis_network.py) which uses plotly also ) and contribute to JOSS, I am unfortunately very busy in the next few months (oversubscribed already). Late June or July is better time for me, but I hope the authors don't have to wait so long.
