### 2019-05-14 #1308 @cMadan
@vc1492a, that sounds good to me. In that case, I'll assign you as reviewer and you can start (and let me know) when you're ready.

@raamana, no worries, thank you for considering!

@jg-you, I'd like to also assign you as reviewer, if you're still interested. I think reviewing alongside another reviewer here would also be helpful in seeing how we do software reviews at JOSS. Let me know if that works for you!

@antoineallard, because of the conflict (co-author as friend), I won't assign you as a reviewer here. Please do look into how reviews work here, and it would be great to have you as a reviewer in the future!  You can sign-up to be on the reviewer list here: http://joss.theoj.org/reviewer-signup.html (@jg-you, please also feel free to add yourself!)
