### 2019-05-19 #1458 @whedon
[REVIEW]: webweb: a tool for creating lightweight network visualizations in the browser
**Submitting author:** @hneutr (<a href="http://orcid.org/0000-0003-2518-841X">Kenneth Hunter Wapman</a>)
**Repository:** <a href="https://github.com/dblarremore/webweb" target ="_blank">https://github.com/dblarremore/webweb</a>
**Version:** 0.0.37
