### 2019-05-23 e01271f
added in 'showLegend' and also decoupled svg generation from saving
