+++ docs/display/meta/showLegend.md
@@ -1,0 +1,8 @@
+---
+name: showLegend
+type: boolean
+default: true
+
+---
+
+If `true`, webweb will display a legend describing how nodes are colored and sized.
