+++ webweb/pygmlion.py
@@ -57,6 +57,7 @@ def get_gml(filename, validate=True):
     formatted = format_gml_dict(content)
     return formatted
 
+
 def get_raw_gml(filename, validate=True):
     with open(filename, 'r') as f:
         lines = [l.strip() for l in f.readlines()]
@@ -71,14 +72,17 @@ def get_raw_gml(filename, validate=True):
 
     return formatted
 
+
 def write_gml(content, filename):
     if type(content) == dict:
         content = convert_gml_dict_to_tuple_list(content)
-    
+
     to_write = [create_element(k, v).write() for k, v in content]
 
     with open(filename, 'w') as f:
         f.write("\n".join(to_write) + "\n")
+
+
 ################################################################################
 #
 #
@@ -100,7 +104,7 @@ def format_gml_dict(value):
         for key, subvalue in value:
             if key == 'comment':
                 continue
-            
+
             if type(subvalue) != list:
                 formatted[key] = subvalue
             else:
@@ -109,8 +113,8 @@ def format_gml_dict(value):
 
                 formatted[key].append(format_gml_dict(subvalue))
 
-
     return formatted
+
 
 def convert_gml_dict_to_tuple_list(content):
     graph_data = content.pop('graph', None)
@@ -131,6 +135,7 @@ def convert_gml_dict_to_tuple_list(content):
 
     return converted
 
+
 def to_tuple_list(item):
     converted = []
     for key, value in item.items():
@@ -143,6 +148,7 @@ def to_tuple_list(item):
             converted.append((key, value))
 
     return converted
+
 
 def create_element(key, value):
     if type(value) is str:
@@ -157,8 +163,9 @@ def create_element(key, value):
     elif type(value) is list:
         return ListAttribute(
             key=key,
-            value=[ create_element(k, v) for k, v in value]
+            value=[create_element(k, v) for k, v in value]
         )
+
 
 ################################################################################
 #
@@ -186,6 +193,7 @@ class GenericAttribute(object):
     def __repr__(self):
         return self.__str__()
 
+
 class LineComment(GenericAttribute):
     def __init__(self, lines=[], key='comment', value=''):
         super().__init__(lines=lines, key=key, value=value)
@@ -193,7 +201,8 @@ class LineComment(GenericAttribute):
     def read(self):
         self.value = self.line
 
-    def is_type(self, lines):
+    @staticmethod
+    def is_type(lines):
         return lines[0].startswith('#')
 
     def __str__(self):
@@ -218,7 +227,8 @@ class KeyValueAttribute(GenericAttribute):
         >>> str(KeyValueAttribute(key='test', value='passes'))
         'test passes (Generic)'
         """
-        return "{content} ({kind})".format(kind=self.kind, content=self.write())
+        return "{content} ({kind})".format(kind=self.kind,
+                                           content=self.write())
 
     def write(self):
         """
@@ -286,6 +296,7 @@ class KeyValueAttribute(GenericAttribute):
         if unconsumed:
             self.unconsumed = [unconsumed.strip()] + self.unconsumed
 
+
 class StringAttribute(KeyValueAttribute):
     """A key/value pair where the value is a quoted string."""
     kind = 'String'
@@ -318,7 +329,7 @@ class StringAttribute(KeyValueAttribute):
 
         if value[0] != '"':
             return False
-        
+
         try:
             self.get_value_and_unconsumed(value)
             return True
@@ -343,7 +354,9 @@ class StringAttribute(KeyValueAttribute):
         >>> StringAttribute(key='test', value='passes').write()
         'test "passes"'
         """
-        return '{key} "{value}"'.format(key=self.key, value=self.value).rstrip()
[…]
