### 2019-05-27 #1458 @vc1492a
**Overview**: Overall, I think the author(s) did a good job on the software and in providing examples and documentation in order to get started easily. The software is easy to use and I think addresses the needs of the audience as well. While I would have liked to see more refinement in the code base prior to consideration for publication in JOSS (e.g. matching and clear format for version numbers, no unused parameters, static methods when needed, etc.), these items can easily be addressed in newer versions of the software prior to acceptance in JOSS. I particularly like the software's flexibility in the input data - being able to use lists, adjacency matrices, or networkx objects is a nice capability and opens up the software to a broader audience versus implementing functionality with networkx alone.

**Version Matching**: The version number submitted to JOSS (0.0.29) does not match the latest version of the source code on the master branch in Github (0.0.30 [as indicated in setup.py](https://github.com/dblarremore/webweb/blob/master/setup.py)) - the JOSS paper and submission will need to be updated for the latest version of the software. Additionally, the version number does not match the [version listed on PyPi](https://pypi.org/project/webweb/) (0.0.32 and the section titled `development` in the project documentation implies a different versioning scheme (e.g. webweb 4, 5) than what is listed on Github and described above, and the source code (`webweb.py`) indicates version 4. It is thus not clear if this project is using [semantic versioning](https://semver.org/) or some other versioning scheme, in addition to the version discrepancies outlined above - these should both be addressed before publication in JOSS. Lastly, I did not see any tagged releases on Github. It may be beneficial to tag the releases that are uploaded to PyPi on Github such that users can download the software directly from Github via a Release as opposed to having to clone or pull the repository.

**Installation Instructions**: The Github repository as well as the linked documentation provide instructions for installation via PyPi or by cloning the Github repository. After examining both `setup.py`, `requirements.txt`, and the source code from the project repository, it seems that the project does not contain any external dependencies but this is in fact not the case as numpy as a requirement for some of the functionality. The project dependencies should be documented in `requirements.txt` and clearly in the documentation and project `readme.md`.

**Functionality**: As indicated in the [pre-review](https://github.com/openjournals/joss-reviews/issues/1308), I am only able to review this software's Python language functionality, specifically Python 3. The functionality works as described by the authors and I particularly like the functionality to view different layers of a network. I did notice a small detail that perhaps could be addressed in future versions of the software. When viewing multiple networks as described in the documentation examples, the viewing options (e.g. node color, size, etc.) are reset when transitioning from one network to another. It would be nice if the previous viewing options were saved in a from of cache so that users do not need to reconstruct a network's view when transitioning from one network to another. This isn't a big deal when using toy examples, but I can imagine some users may be frustrated if constructing a particular view in network A, transitioning to network B, only to find that the view is gone when transitioning back to network A. The ability to zoom into and out of a visualized network using the mouse or trackpad would be a welcome addition as well. Lastly, while the `Highlight nodes named` functionality works well, I think it may be more intuitive to place the `Show node names`  checkbox above the free text field used to highlight particular nodes. I did not test for functionality in Python 2 - if Python 2 is not supported, this should be made explicit in `readme.md` or in the documentation.

I noticed a lot of use of Python's `os` library. While this works fine on Linux and Unix devices, it can result in errors on Windows machines. Using `Pathlib` in place of `os` would reduce the likelihood of errors on Windows. I am not able to verify the functionality of this software on Windows machines (the editor can decide whether this functionality on windows should be formally verified or whether this potential limitation should be listed in `readme.md` or documentation), but do know that `Pathlib` would allow the developers to normalize paths for Windows machines easily. Consider opening an issue with the tag `enhancement`.

Lastly, the GUI in webweb seems to share a lot of similar capabilities and features with the GUI used in [netwulf](https://github.com/benmaier/netwulf). Although the aims of each software are different, both sets of authors could potentially reduce future development effort by collaborating on shared features and/or components.

**Automated Tests**: While I did see a `test.py` file which I ran successfully, the code contained therin could only be used to verify some of the functionality of webweb. Additionally, I did not see any form of automated tests. While writing automated tests for software that requires some form of user interaction presents its own unique challenges, revisions of this work should include a series of unit tests that can be used to verify the functionality of the software, through testing of function outputs, input and output formatting, etc. There are a variety of ways to write and use unit tests in Python, two of which are the `pytest` and `pytest-cov` packages - the former allows developers to execute unit tests (with helpful extensions for writing tests, too) while the latter provides functionality to `pytest` for coverage statistics, e.g. which lines of code your unit tests cover and the percentage of lines in each file (and overall, the package) that is covered by the written unit tests (this coverage statistic could be reported in `readme.md`). Additionally, if the authors looking to have others contribute to the project more easily in the future consider setting up a continuous integration pipeline that will automatically run tests and report the test results (e.g. using Travis CI) and coverage (e.g. using Coveralls) with any pushes to `master` or `dev` and for any pull requests opened by open-source contributors. This pipeline would ease the burden on making sure new additions and/or changes to software work as intended, and would also allow maintainers to formally test against every version of Python 3 that should be supported. This would allow the authors to state "this software works in Python versions 3.5 - 3.7" as opposed to simply stating "works in Python 3". In preparation for this, I have placed the `test.py` and related files in a directory called `tests` in my [pull request](https://github.com/dblarremore/webweb/pull/39).

**Example Usage**: The software documentation does include some examples on how to use the software, which I found helpful for reviewing the software's functionality. I also saw a section for real-world examples which is great!

**References**: `paper.bib` had all relevant references, but did not have all DOIs that are available for these references, so added a few and opened a [this pull request](https://github.com/dblarremore/webweb/pull/39). Specifically for the `netwulf` reference, the DOI is pending since it's currently under review by JOSS (https://github.com/openjournals/joss-reviews/issues/1425). The editor of this work should decide how to handle this reference and whether updating the `paper.bib` file post publication in JOSS is a possibility so that this reference's DOI may be included for this work once itself is reviewed and later published.

**Community Guidelines**: The Github repository as well as the documentation contains some guidelines on how to report issues and how to contribute to webweb's development which is great! It may benefit the authors / developers however to include more specific contribution instructions, such as asking users to first fork the repository and then create a new branch with a name that describes the issue or new feature before submitting a pull request. This may result in some additional organization. Additionally, while there is a section titled `development` in the documentation that serves as a historical record of software updates, it would be nice to include a `changelog.md` file in the project repository so that users and/or developers do not need to visit the documentation in order to see a record of project updates.

**Other**: I made a series of other small updates which are contained in the aforementioned pull request, such as moving imports to the top of the files, converting functions that don't use object attributes and functions to static methods, adding numpy to `requirements.txt`, and removing unused parameters from some functions. Additionally, while the paper does provide some acknowledgements, it would be nice to include the affiliations for those acknowledged in this section, if available.
