+++ changelog.md
@@ -1,0 +1,65 @@
+# 20190526 - 0.0.32
+
+- added `display.showLegend` which defaults to = `True`. If `True`, will draw the legend.
+
+# 20190206 - canvas!
+
+- webweb now renders with canvas! It’s faster and able to support larger networks.
+- Legends are labeled with what they make legendary.
+
+# 20190130 - 0.0.5
+
+- Documentation is live!
+- webweb is on pip!
+- matlab is completely rewritten!
+- webweb visualizations have no dependencies
+- drag&drop and json saving removed
+- all node names can be shown
+- interactive menus can be hidden
+- binary colors and sizes can be inverted
+- nodes can be strings (python)
+- webweb is responsive (mostly)
+
+# 20181125 - 0.0.4
+
+Merged fork from Hunter Wapman
+
+- supports multi-layer networks
+- all display parameters can be set
+- saveable json
+- color palette can be changed
+
+# 20180123 - 0.0.3.4
+
+Merged a fork from Michael Iuzzolino
+
+- SVG saving
+- drag and drop file loading
+- standard file loading.
+
+# 20150522 - 0.0.3.3
+
+- pause/play node movement
+
+# 20140516 - 0.0.3.2
+
+- link opacity defaults to checked
+- charge selection bugfix
+- link strength can be changed
+- webweb now runs on Windows
+
+# 20131222 - 0.0.3.1
+
+- accepts a single sparse matrix as input
+
+# 20131010 - webweb is featured in a publication!
+
+in [PLoS Computational Biology](http://www.ploscompbiol.org/article/info:doi/10.1371/journal.pcbi.1003268), along with [interactive versions of all published figures](http://danlarremore.com/var).
+
+# 20130826 - 0.0.3
+
+a new and improved version of webweb!
+
+# 2013087 - 0.0.3
+
+a new and improved version of webweb!
