### 2019-06-04 b8ded56
changing to a smaller font for changelog
