+++ changelog.md
@@ -1,13 +1,13 @@
-# 20190526 - 0.0.32
+## 20190526 - 0.0.32
 
 - added `display.showLegend` which defaults to = `True`. If `True`, will draw the legend.
 
-# 20190206 - canvas!
+## 20190206 - canvas!
 
 - webweb now renders with canvas! It’s faster and able to support larger networks.
 - Legends are labeled with what they make legendary.
 
-# 20190130 - 0.0.5
+## 20190130 - 0.0.5
 
 - Documentation is live!
 - webweb is on pip!
@@ -20,7 +20,7 @@
 - nodes can be strings (python)
 - webweb is responsive (mostly)
 
-# 20181125 - 0.0.4
+## 20181125 - 0.0.4
 
 Merged fork from Hunter Wapman
 
@@ -29,7 +29,7 @@ Merged fork from Hunter Wapman
 - saveable json
 - color palette can be changed
 
-# 20180123 - 0.0.3.4
+## 20180123 - 0.0.3.4
 
 Merged a fork from Michael Iuzzolino
 
@@ -37,29 +37,29 @@ Merged a fork from Michael Iuzzolino
 - drag and drop file loading
 - standard file loading.
 
-# 20150522 - 0.0.3.3
+## 20150522 - 0.0.3.3
 
 - pause/play node movement
 
-# 20140516 - 0.0.3.2
+## 20140516 - 0.0.3.2
 
 - link opacity defaults to checked
 - charge selection bugfix
 - link strength can be changed
 - webweb now runs on Windows
 
-# 20131222 - 0.0.3.1
+## 20131222 - 0.0.3.1
 
 - accepts a single sparse matrix as input
 
-# 20131010 - webweb is featured in a publication!
+## 20131010 - webweb is featured in a publication!
 
 in [PLoS Computational Biology](http://www.ploscompbiol.org/article/info:doi/10.1371/journal.pcbi.1003268), along with [interactive versions of all published figures](http://danlarremore.com/var).
 
-# 20130826 - 0.0.3
+## 20130826 - 0.0.3
 
 a new and improved version of webweb!
 
-# 2013087 - 0.0.3
+## 2013087 - 0.0.3
 
 a new and improved version of webweb!
