### 2019-06-05 1123545
adding automated testing for python; changing versioning; slight shift of temperature for node name display
