### 2019-06-05 dblarremore/webweb#40 @hneutr
Add a pipeline for continuous integration
This will make contributing easier, and we should do it.
