### 2019-06-05 dblarremore/webweb#41 @hneutr
Cache display settings when changing networks
> When viewing multiple networks as described in the documentation examples, the viewing options (e.g. node color, size, etc.) are reset when transitioning from one network to another. It would be nice if the previous viewing options were saved in a from of cache so that users do not need to reconstruct a network's view when transitioning from one network to another.

- vc1492a
