### 2019-06-05 dblarremore/webweb#42 @hneutr
Tag PyPi releases
This will allow users to get the latest version of the code without cloning the repository.
