### 2019-06-05 #1458 @hneutr
> Version Matching: The version number submitted to JOSS (0.0.29) does not match the latest version of the source code on the master branch in Github (0.0.30 as indicated in setup.py) - the JOSS paper and submission will need to be updated for the latest version of the software.

I've updated `setup.py` and unified the versioning scheme to match what is on PyPi. I've updated (after making — see point below) the changelog to reflect this versioning scheme.

I apologize if I'm missing where this is described, but I'm not sure how to update the JOSS submission.

> Lastly, I did not see any tagged releases on Github. It may be beneficial to tag the releases that are uploaded to PyPi on Github such that users can download the software directly from Github via a Release as opposed to having to clone or pull the repository.

I agree this would be useful. I'm unfamiliar with the tagging process, but am reading up now, and have created [an issue](https://github.com/dblarremore/webweb/issues/42) to represent this.

> The project dependencies should be documented in requirements.txt and clearly in the documentation and project readme.md.

Done. I've updated the readme to point out the requirement of numpy, as well as clarified the index page of the documentation site to make clear that the _visualizations_ have no dependencies, as opposed to the software used to generate those visualizations. Great point!

> I did notice a small detail that perhaps could be addressed in future versions of the software. When viewing multiple networks as described in the documentation examples, the viewing options (e.g. node color, size, etc.) are reset when transitioning from one network to another. It would be nice if the previous viewing options were saved in a from of cache so that users do not need to reconstruct a network's view when transitioning from one network to another.

I absolutely agree. I think this pretty annoying. I've added a [feature request](https://github.com/dblarremore/webweb/issues/41) to represent this, and quoted you as I think you convey the issue and solution well.

> The ability to zoom into and out of a visualized network using the mouse or trackpad would be a welcome addition as well.

There's an [issue](https://github.com/dblarremore/webweb/issues/18) for this — I agree that it would be really useful!

> Lastly, while the Highlight nodes named functionality works well, I think it may be more intuitive to place the Show node names checkbox above the free text field used to highlight particular nodes.

No strong feelings to putting it above as opposed to below, but also not entirely sure how big a difference it would make. My general feeling is "if it ain't broke, don't fix it", especially with UI, but I'm certainly open to being convinced of otherwise. What draws you wanting it above?

> If Python 2 is not supported, this should be made explicit in readme.md or in the documentation.

Added an explicit disclaimer that the code may not function in python 2.

> Using Pathlib in place of os would reduce the likelihood of errors on Windows.

`os` has been removed in favor of Pathlib. Thanks for pointing me in the direction — this is definitely an improvement!

> The GUI in webweb seems to share a lot of similar capabilities and features with the GUI used in netwulf.

I've talked to Dan, and I think collaborating with them is a great idea, and will be reaching out.

> Automated Tests: Additionally, I did not see any form of automated tests. While writing automated tests for software that requires some form of user interaction presents its own unique challenges, revisions of this work should include a series of unit tests that can be used to verify the functionality of the software, through testing of function outputs, input and output formatting, etc.

I've added a bunch of automated tests that `pytest` will run! This covers 70% of the code.

> If the authors looking to have others contribute to the project more easily in the future consider setting up a continuous integration pipeline that will automatically run tests and report the test results (e.g. using Travis CI) and coverage (e.g. using Coveralls) with any pushes to master or dev and for any pull requests opened by open-source contributors.

I agree that this would be desirable, and have opened [an issue](https://github.com/dblarremore/webweb/issues/40) to represent this.

> References: paper.bib had all relevant references, but did not have all DOIs that are available for these references, so added a few and opened a this pull request.

I've merged the pull request. Thank you for taking the time to add the DOIs!

> Specifically for the netwulf reference, the DOI is pending since it's currently under review by JOSS (#1425). The editor of this work should decide how to handle this reference and whether updating the paper.bib file post publication in JOSS is a possibility so that this reference's DOI may be included for this work once itself is reviewed and later published.

@cmadan, is this something we can do?

> It may benefit the authors / developers however to include more specific contribution instructions, such as asking users to first fork the repository and then create a new branch with a name that describes the issue or new feature before submitting a pull request.

Done!

> it would be nice to include a changelog.md file in the project repository so that users and/or developers do not need to visit the documentation in order to see a record of project updates.

I added a `changelog.md` while to the github repo and replaced the `development` section with the same — these will automatically be kept in-sync, too. Thanks for the tip!

> Other: I made a series of other small updates which are contained in the aforementioned pull request

Merged! Thank you again for taking the time to do this.

> while the paper does provide some acknowledgements, it would be nice to include the affiliations for those acknowledged in this section

Added!
