### 2019-06-05 #1458 @hneutr
@vc1492a thank you for your thorough review! I believe I've addressed your points — they were all excellent, and improved the software. The perspective of a seasoned python developer is much appreciated!

@cMadan I've responded to @vc1492a's review above!

I believe I hit everything mentioned, but if there's any points I've missed, please let me know and I'll circle back.
