### 2019-06-06 6e26fd1
fixing bug (gh issue 43)
