### 2019-06-07 #1458 @vc1492a
@hneutr Thanks for your detailed response! I've examined the updates and they look good to me. In reply to your question about what drives me to prefer that ordering on the UI, it just seems to be that a user would first toggle on the option to show node names before doing a secondary filtering.
