### 2019-06-10 dblarremore/webweb#44 @jg-you
Switch to pathlib broke webweb show() on linux
I'm currently testing the latest version of webweb (0.0.35) for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1458).

Here's the new output of the minimal code:

    In [1]: import webweb

    In [2]: from webweb import Web
       ...:
       ...: # make a list of unweighted edges
       ...: edge_list = [[1, 2], [2, 3], [3, 4]]
       ...:
       ...: # instantiate webweb and show the result
       ...: Web(edge_list).show()

    In [3]: gio: file://tmp/webweb.html: Error when getting information for file “/webweb.html”: No such file or directory

I'm running archlinux, python 3.7.3, webweb 0.0.35
