### 2019-06-10 dblarremore/webweb#44 @jg-you
As far as *Linux* is concerned, removing "file:/"  [from line 90 of webweb.py](https://github.com/dblarremore/webweb/blob/master/webweb/webweb.py#L90) should fix the bug.
Don't know how this affects Windows / Mac?
