### 2019-06-11 dblarremore/webweb#44 @hneutr
I appreciate you pointing this out, and for the triage — the information is really helpful.

I tested the change and, on osx at least, the removal causes the file to not be opened. I am looking into how to resolve this. I’ll post again when I’ve got a working solution!
