### 2019-06-11 dblarremore/webweb#44 @hneutr
@jg-you, after a bit of exploring, it seems (to me) there are only two places that could be causing this:
- switching `file://` --> `file:/`
- using `path.resolve()`, though I'm not sure why that would.

I'm proposing changing the line to the following:
`webbrowser.open_new("file://" + str(path))`
which I've confirmed works on mac, and believe addresses both possible issues, though I haven't been able to test it on a linux environment.

Would you be willing to try subbing that line in, and seeing if that works?
