### 2019-06-11 #1458 @jg-you
`webweb` is a neat little package that definitely addresses a need of the Network Science community.
It sets out to do one thing---lightweight and easy graph interactive graph viz---and it does it well, so I'm overall pretty excited about the project.

I have been using 0.0.18 in my work for a while, and updated to version 0.0.35, the latest release on PyPI, for this review.
I could not test the MATLAB interface either, because it is incompatible with `octave` (due to the use of `jsonencode`), the open-source alternative to which I have access.

Overall, the package passes almost all the checks with very few exceptions.
The review of @vc1492a covers most of the issues I could identify, so my report will be very short.

**Version numbers**: I'll second @vc1492a in saying that tagging releases on github would be nice, for a different reason:
PyPI does not send out update notifications, but it is easy to watch a github repository for releases.
As the maintainer of the [AUR package for python-webweb](https://aur.archlinux.org/packages/python-webweb/), it would help me not let the package become out-of-date (as stated above: I was still on v0.0.18!).

**Documentation**: The examples listed on the website are fantastic for getting an intuitive grasp of what the package is doing.
If I have one comment, it is that the docstring of the all-important `Web` object could be improved.
It is hard to get a meaningful description of what `Web` does without going through the online examples because (1) there are no example nor exhaustive description of the interface in the docstring, and (2) the signature doesn't reflect its functionality (nx_G is passed as a kwargs to ``Network`` for example).
This is a minor pain point, but since `Web` is the main interface with the user, taking the time to document it better would improve their experience.
This is especially useful for users (like me!) that always forget what is the exact way to call a module ;)

**Installation**: The numpy dependency now appears everywhere on github :+1:, but it could also be added to the installation section of [the webweb website](https://webwebpage.github.io/).

----

What is below is not part of my review per se, but just little comments.

**Possible features?**: Some things that would be great to have at some point.
1. Multiple nodes highlight in the GUI (say, with a comma separated list). As far as I can tell, the user can only highlight one node at a time in the current build.
2. CLI interface. When I want to visualize a small network file, I'll typically boot up an ipython instance, import `networkx` and `webweb`, then I'll load the network and I'll finally visualize it. A neat improvement could be to add a `main` to webweb and make it callable from the CLI directly, e.g. `ẁebweb network.edgelist` would load `network.edgelist` and open the browser to the webweb page.
3. Use an open-source alternative to `jsonencode` in the MATLAB interface to make `webweb` accessible to octave users.

**Minor things**: Consider maintaining the author information in `__version__`, `__authors__`, etc., dunders string.  Consider adding a hash to the webweb.html page so that many instance can be ran and refreshed without overwriting one another.
