### 2019-06-17 dblarremore/webweb#48 @hneutr
allow highlighting of multiple nodes in the `matchNodesNamed` widget
probably by comma separating them. it would be pretty easy to do!
