### 2019-06-17 dblarremore/webweb#49 @hneutr
add a CLI for webweb
a simple command like `webweb [edgelist file here]` would be _awesome_.

More features to be added would be perhaps giving node and metadata.
