+++ webweb/webweb.py
@@ -14,8 +14,11 @@ import webbrowser
 import tempfile
 import numpy as np
 from pathlib import Path
+import sys
 
-from . import pygmlion
+sys.path.append(str(Path(__file__).parent))
+
+import pygmlion
 
 
 class Web(dict):
@@ -23,20 +26,80 @@ class Web(dict):
     a collection of named webweb Network objects, a set of display parameters, and a title
     """
 
-    def __init__(self, adjacency=[], title="webweb", display={}, **kwargs):
-        """parameters:
-        - adjacency: adjacency to make a visulization from. see `Network.add_layer`
+    def __init__(
+            self,
+            adjacency=[],
+            title="webweb",
+            display={},
+            adjacency_type=None,
+            nodes={},
+            metadata=None,
+            nx_G=None,
+            gml_file=None,
+    ):
+        """
+        usage:
+        - `web = Web(adjacency=[[0, 1]], title='a web woohoo')`
+
+        parameters:
+        - adjacency: edge list or adjacency matrix
         - title: string. Will set the html title of the visualization if display.attachWebwebToElementWithId = None
         - display: dictionary of display parameters
-        - see `Network.add_layer` for all other parameters."""
+        - adjacency_type: string. 'matrix' or 'edge list'
+        - nodes: dict of node attribute dicts
+        ```json
+        {
+            'key1': {
+                'attribute1': 'value1',
+                ...
+            },
+            ...
+        }
+        ```
+        - metadata: dict of vectorized metadata and display information.
+        ```python
+        {
+            'attribute': {
+                'values': [ "attribute_value", ...],
+
+                # `categories` only needs to be supplied if `values` holds
+                # categorical data that is represented by numbers.
+                # the values in the `values` array will be used as indexes to
+                # this array.
+                'categories': ["category1", "category2", ...]
+
+                # `type` only needs to be set if you're displaying binary
+                # information with 0/1 instead of True/False
+                'type': 'binary',
+            }
+        }
+        ```
+        - nx_G: a networkx graph.
+        - gml_file: path to a gml file
+
+        ---
+
+        nodes which appear in both the adjacency and as keys in the `nodes`
+        dictionary will be given the values of the attributes under their
+        corresponding key in the `nodes` dictionary
+
+        `adjacency_type` only needs to be supplied if your adjacency is
+        represented as a matrix and that matrix has 3 or fewer nodes
+        """
         self.title = title
         self.display = Display(display)
         self.networks = Networks()
 
         # if we have an adjacency, add it into the networks object
-        if len(adjacency) or kwargs.get('nx_G') or kwargs.get('gml_file'):
-            kwargs['adjacency'] = adjacency
-            getattr(self.networks, self.title)(**kwargs)
+        if len(adjacency) or nx_G or gml_file:
+            getattr(self.networks, self.title)(
+                adjacency=adjacency,
+                adjacency_type=adjacency_type,
+                nodes=nodes,
+                metadata=metadata,
+                nx_G=nx_G,
+                gml_file=gml_file,
+            )
 
     @staticmethod
     def base_path():
@@ -201,8 +264,15 @@ class Network(dict):
         elif len(kwargs.keys()):
             self.add_layer(**kwargs)
 
-    def add_layer(self, adjacency=[], adjacency_type=None, nodes={},
-                  metadata=None, nx_G=None, gml_file=None):
+    def add_layer(
+            self,
+            adjacency=[],
+            adjacency_type=None,
+            nodes={},
+            metadata=None,
+            nx_G=None,
+            gml_file=None
+    ):
         """adds a layer to the network.
 
         parameters:
