### 2019-06-18 12bcc73
updating docstrings, adding hashs to file paths
