+++ setup.py
@@ -20,8 +20,8 @@ def read_requirements(filename):
 
 setup(
     name="webweb",
-    version="0.0.36",
-    author="Dan Larremore; Hunter Wapman",
+    version="0.0.37",
+    author="Daniel Larremore; Hunter Wapman",
     author_email="hunter.wapman@gmail.com",
     description="webweb: an easy-to-use network visualization tool",
     long_description="webweb: an easy-to-use network visualization tool",
