### 2019-06-18 dblarremore/webweb#51 @hneutr
use an open source alternative to `jsonencode` in matlab version
so that `octave` users can run the code!
