### 2019-06-18 #1458 @hneutr
Hey @jg-you! Thank you for the thorough review, and the great suggestions. Responding inline below!

> tagging releases on github would be nice

Upping this priority!

> the docstring of the all-important Web object could be improved.
> (1) there are no example nor exhaustive description of the interface in the docstring
> (2) the signature doesn't reflect its functionality

done in `0.0.37`!

> The numpy dependency could be added to the installation section of the webweb website.

added!

> Multiple nodes highlight in the GUI (say, with a comma separated list)

Made [an issue](https://github.com/dblarremore/webweb/issues/48)!

> CLI interface. A neat improvement could be to add a main to webweb and make it callable from the CLI directly, e.g. ẁebweb network.edgelist would load network.edgelist and open the browser to the webweb page.

I think this is an _awesome_ idea, very much inline with the "spirit" of `webweb`. [Issue here](https://github.com/dblarremore/webweb/issues/49).

> Use an open-source alternative to jsonencode in the MATLAB interface to make webweb accessible to octave users.

Made an [issue](https://github.com/dblarremore/webweb/issues/51)

> Consider maintaining the author information in __version__, __authors__, etc., dunders string.

done in `0.0.37`!

> adding a hash to the webweb.html page so that many instance can be ran and refreshed without overwriting one another.

done in `0.0.37`!
