### 2019-06-21 a349d07
Dunder in init
