+++ webweb/__init__.py
@@ -1,3 +1,6 @@
 name = "webweb"
 
+__version__ = '0.0.37'
+__authors__ = 'Daniel Larremore; Hunter Wapman'
+
 from .webweb import Web
