### 2019-06-21 #1458 @jg-you
@cMadan: @hneutr applied all the changes (and more) highlighted in my review.
