### 2019-07-02 #1458 @cMadan
@vc1492a @jg-you, thank you both for your thorough reviews, I greatly appreciate it!

@hneutr, it looks like you're still not using GitHub releases? (And I don't see a response to that suggestion.)

The netwulf DOI will be 10.21105/joss.01425. I am fine with you citing it here, with this DOI (even though not officially registered yet) as it will resolve as a citation to the other package once it is published.

Otherwise I think the review is nearly done and I just need to do some final checks.
