### 2019-07-08 b502da8
updating paper.bib for netwulf doi
