+++ paper/paper.bib
@@ -54,7 +54,8 @@
     title = {netwulf: Interactive visualization of networks based on Ulf Aslak's d3 web app},
     url = {https://github.com/benmaier/netwulf},
     year = {2018},
-    month = {November}
+    month = {November},
+    doi = {10.21105/joss.01425}
 }
 
 @article{PhysRevE.90.012805,
