### 2019-07-08 #1458 @hneutr
@cMadan:

I've added the netwulf DOI to the `paper.bib` file. I've also tagged the most recent release on GitHub, and will continue to tag commits that are pushed to PyPi.

Let me know if there is anything else that needs to be done!
