+++ client/js/network.js
@@ -1,0 +1,460 @@
+import * as d3 from 'd3'
+import { AllSettings } from './controller'
+import { NodeKey } from './node'
+import { Link } from './link'
+export class Network {
+  // TODO:
+  // - add in the idea of a network knowing about the metadata it has that are relevant
+  //    - i.e., should _here_ call `getMetadatumType`
+
+  // standardizes the format of a network
+  constructor(name, data, globalMetadata, globalNodes) {
+    this.name = name
+    this.metadata = {}
+    this.globalMetadata = globalMetadata
+    this.globalNodes = globalNodes || {};
+
+    this.layers = this.makeLayers(data)
+
+  }
+
+  updateState(metadata) {
+    this.state = new AllSettings(metadata)
+  }
+
+  // if there are no layers, put things into layers
+  makeLayers(data) {
+    let layers = data.layers
+
+    if (layers == undefined) {
+      layers = [
+        {
+          'edgeList' : data.edgeList,
+          'nodes' : data.nodes,
+          'metadata' : data.metadata,
+          'display': data.display,
+        }
+      ]
+    }
+
+    for (let [i, layer] of layers.entries()) {
+      // edgelist operations
+      let edgeList = layer.edgeList == undefined ? [] : layer.edgeList
+      edgeList = this.regularizeLayerEdgeList(edgeList)
+
+      // metadata operstions
+      let metadata = layer.metadata == undefined ? {} : layer.metadata
+      metadata = this.regularizeLayerMetadata(metadata, this.globalMetadata)
+
+      // node operstions
+      let nodes = layer.nodes == undefined ? {} : layer.nodes
+      nodes = this.regularizeLayerNodes(nodes, this.globalNodes)
+
+      let nodeNameToIdMap = this.getNodeNameToIdMap(edgeList, metadata, nodes)
+      let nodeIdToNameMap = Object.fromEntries(Object.entries(nodeNameToIdMap).map(([k, v]) => ([v, k])))
+
+      // assign metadata from the metadata vector onto the nodes
+      nodes = this.mergeMetadataWithNodes(nodes, metadata, nodeIdToNameMap)
+
+      // TODO:
+      // console.log(JSON.stringify('must do something here with edgeWeights'));
+      // this.scales.links.width.domain(d3.extent(edgeWeights));
+      // this.scales.links.opacity.domain(d3.extent(edgeWeights));
+      let links = this.createLinks(edgeList, nodeNameToIdMap)
+
+      nodes = this.addEdgeMetadataToNodes(links, nodes, this.isUnweighted(nodes))
+
+      let displayableMetadata = this.getDisplayableMetadata(nodes, metadata)
+
+      layers[i] = {
+        "edgeList": edgeList,
+        "links": links,
+        "displayableMetadata":displayableMetadata,
+        "metadata": metadata,
+        "nodes": nodes,
+        "nodeNameToIdMap": nodeNameToIdMap,
+        "nodeIdToNameMap": nodeIdToNameMap,
+        "edgeWeights": links.map(link => link.weight),
+        "nodeCount": Object.keys(nodeNameToIdMap).length,
+        'display': layer.display,
+      }
+    }
+
+    return layers
+  }
+
+  // iterates over all of the nodes and identifies the set of metadata we can
+  // include in the visualization.
+  getDisplayableMetadata(nodes, metadata) {
+    let allMetadata = {
+      'degree': {
+        'type': 'degree',
+      },
+      'strength': {
+        'type': 'degree',
+      },
+    }
+
+    let nodeKeyer = new NodeKey()
+
+    Object.entries(nodes).forEach(([id, node]) => {
+      let metadataKeys = nodeKeyer.filterMetadataKeys(Object.keys(node), node)
[…]
