+++ client/js/node.js
@@ -1,7 +1,8 @@
 import * as d3 from 'd3'
+import { Text } from './text'
 
 export class NodeKey {
-  get nonMetadataKeys() {
+  static get nonMetadataKeys() {
     return [
       'degree',
       'strength',
@@ -13,14 +14,16 @@ export class NodeKey {
       'vy',
       'x',
       'y',
+      'nodeKeyer',
+      'settings',
     ]
   }
 
-  isNonMetadataKey(key) {
+  static isNonMetadataKey(key) {
     return this.nonMetadataKeys.indexOf(key) >= 0 ? true : false
   }
 
-  isMetadataKey(key, obj) {
+  static isMetadataKey(key, obj) {
     if (this.isNonMetadataKey(key)) {
       return false
     }
@@ -38,63 +41,26 @@ export class NodeKey {
     return false
   }
 
-  filterMetadataKeys(keys, obj) {
+  static filterMetadataKeys(keys, obj) {
     return keys.filter((key) => this.isMetadataKey(key, obj), this)
-  }
-
-  setMetadataOntoObject(key, obj, metadata) {
-    let type = metadata.type
-    if (type !== undefined) {
-      if (type == 'categorical') {
-        if (metadata.categories !== undefined) {
-          var nodeCategoryNumber = obj[key]
-          if (isInt(nodeCategoryNumber)) {
-            obj[key] = metadatum.categories[nodeCategoryNumber];
-          }
-        }
-      }
-      else if (type == 'binary') {
-        obj[key] = obj[key] ? true : false;
-      }
-    }
   }
 }
 
 
 export class Node {
-  constructor(idx) {
+  constructor(idx, settings) {
     this.idx = idx
-    this.settings = null
-    this.nodeKeyer = NodeKey()
-  }
-
-  set settings(settings) {
     this.settings = settings
   }
 
-  setMetadata() {
-    for (let key of nodeKeyer.filterMetadataKeys(Object.keys(this), this)) {
-      if (this.settings.metadata !== undefined && this.display.metadata[key] !== undefined) {
-        let metadatumInfo = this.display.metadata[key]
-
-        let value = this[key]
-
-        // if we have categories, change the node values here
-        if (metadatumInfo.categories !== undefined) {
-          if (isInt(value)) {
-            var nodeCategoryNumber = this.nodes[i][key];
-            this[key] = metadatumInfo.categories[nodeCategoryNumber];
-          }
-        }
-        else if (metadatumInfo.type !== undefined && metadatumInfo.type == 'binary') {
-          this[key] = value ? true : false;
-        }
-      }
+  setMetadata(metadata) {
+    for (let [key, value] of Object.entries(metadata)) {
+      this.key = value
     }
   }
 
   resetMetadata() {
-    this.metadataKeys.forEach((key) => {
+    NodeKey.filterMetadataKeys(Object.keys(this), this).forEach((key) => {
       delete this[key]
     }, this)
 
@@ -104,29 +70,29 @@ export class Node {
 
   get radius() {
     if (this.fixedRadius) {
-      return this.fixedRadius;
+      return this.fixedRadius
     }
 
-    let radius = this.__scaledSize * this.settings.r;
+    let radius = this.__scaledSize * this.settings.r || this.settings.r
     if (this.matchesString() || this.containsMouse(radius)) {
       radius *= 1.3;
     }
 
-    return radius;
+    return radius
   }
 
   matchesString() {
-    let matchString = this.settings.nameToMatch;
+    let matchString = this.settings.nameToMatch
     if (matchString !== undefined && matchString.length > 0) {
[…]
