+++ client/js/__tests__/layer_spec.js
@@ -1,0 +1,619 @@
+import { Layer } from '../layer'
+
+describe("layer object metadata regularization", () => {
+  const globalMetadata = {
+    'test': {
+      'categories': ["this", 'is', 'a', 'test'],
+      'type': 'categorical'
+    }
+  }
+  const globalNodes = {
+    1: {
+      'name': 1,
+    },
+    2: {
+      'name': 2,
+    },
+  }
+  let layer = new Layer([], {}, {}, {}, globalMetadata, {})
+
+  it("tests that global metadata is defaulted", () => {
+    let actual = layer.regularizeMetadata({}, globalMetadata)
+    
+    expect(actual).toStrictEqual(globalMetadata)
+  })
+
+  it("tests that layer metadata is applied", () => {
+    let actual = layer.regularizeMetadata({
+      'test': {
+        'type': 'testtype',
+      },
+      'other_test': {
+        'categories': ['A', 'B', 'C'],
+        'values': [1,2,3],
+      },
+    }, globalMetadata)
+    
+    expect(actual).toStrictEqual({
+      'test': {
+        'categories': ["this", 'is', 'a', 'test'],
+        'type': 'testtype'
+      },
+      'other_test': {
+        'categories': ['A', 'B', 'C'],
+        'values': [1,2,3],
+      }
+    })
+  })
+})
+
+describe("tests layer object nodes regularization", () => {
+  const globalMetadata = {}
+
+  let nodes = {
+    1: {
+      'hi': 1,
+    },
+    2: {
+      'hi': 2,
+    },
+    'non-int': {
+      'hi': 3,
+    }
+  }
+
+  it("tests that nodes are properly formed", () => {
+    let layer = new Layer([], nodes, {}, {}, globalMetadata, {})
+    let actual = layer.regularizeNodes(nodes, {})
+    
+    expect(actual).toStrictEqual({
+      1: {
+        "degree": 0,
+        'hi': 1,
+        "name": "1",
+      },
+      2: {
+        "degree": 0,
+        'hi': 2,
+        "name": "2",
+      },
+      '1': {
+        "degree": 0,
+        'hi': 1,
+        "name": "1",
+      },
+      '2': {
+        "degree": 0,
+        'hi': 2,
+        "name": "2",
+      },
+      'non-int': {
+        "degree": 0,
+        'hi': 3,
+        "name": "non-int",
+      }
+    })
+  })
+
+  it("tests that global node metadata is added to nodes, but does not overwrite", () => {
+    let globalNodes = {
+      1: {
+        'hi': 0,
+        'name': 1,
+      },
+      2: {
+        'hi': 0,
+        'name': 2,
+      },
+    }
+
+    let layer = new Layer([], nodes, {}, {}, globalMetadata, globalNodes)
+    let actual = layer.regularizeNodes(nodes, globalNodes)
+    
+    expect(actual).toStrictEqual({
+      1: {
+        "degree": 0,
+        "strength": 0,
+        'hi': 1,
+        'name': "1",
+      },
+      2: {
+        "degree": 0,
+        "strength": 0,
+        'hi': 2,
+        'name': "2",
+      },
+      'non-int': {
+        "degree": 0,
+        "strength": 0,
+        'hi': 3,
+        'name': 'non-int',
+      }
+    })
+  })
+})
+
+describe("tests that nodes are pulled from their objects properly", () => {
+  let layer = new Layer([], {}, {}, {}, {}, {})
+  it("tests that edges come out as sorted when all numbers", () => {
+    let edges = [[2, 1]]
+    let nodes = {
+      4: {},
+      3: {},
+    }
+    let actual = layer.getNodesMentioned(edges, {}, nodes)
[…]
