### 2019-08-07 14417ab
getting pretty close. have node names mostly, and switching.
