+++ client/js/attribute.js
@@ -1,0 +1,369 @@
+import * as d3 from 'd3'
+
+/********************************************************************************
+ * Attribute object
+ *
+ * handles:
+ * - in IO:
+ *    - reading in of data in different formats
+ *    - guessing what type attribute things are
+ *    - transforming the input when "mapped categorical" type
+ * -  on 'updateScales':
+ *    - getting the raw and scaled values to color and size by
+ *    - slapping them onto the nodes
+********************************************************************************/
+export class Attribute {
+  static TYPE = 'none'
+
+  static defaultColorValue = d3.rgb(100, 100, 100)
+  static defaultValue = 1
+
+  static DISPLAY_COLOR = true
+  static DISPLAY_SIZE = true
+
+  get sizeScale() {
+    return 'nodeSize'
+  }
+
+  // want to opaquly handle color and size
+  // also want 'shouldDisplay'
+  constructor(key) {
+    this.key = key
+  }
+
+  get nodes() {
+    return this._nodes
+  }
+
+  set nodes(nodes) {
+    this._nodes = nodes
+  }
+
+  static typeIsStatedType(type) {
+    if (type !== undefined && type == this.TYPE) {
+      return true
+    }
+    return false
+  }
+
+  // EXTENT and other size properties
+  extent(nodes) {
+    let nodeValues = Object.values(nodes).map(node => node[this.key])
+    return d3.set(nodeValues).values().sort()
+  }
+
+  colorExtent(nodes) {
+    return d3.set(this.getRawColorValues(nodes)).values().sort()
+  }
+
+  valueSetSize(nodes) {
+    return this.extent(nodes).length
+  }
+
+  /*
+   * Parameters:
+   * - nodes
+   *
+   * Returns: 
+   *    boolean. True if the values are of the type
+    * */
+  isType(nodes) {
+    return false
+  }
+
+
+  /********************************************************************************
+   * Transformation Functions
+   *
+   * these modify the imput, and run on dictionaries of elements. 
+   * run as part of the startup process. 
+   * stored in layers after.
+  ********************************************************************************/
+
+  /*
+   * Parameters:
+   * - node
+   * 
+   * Returns:
+   *    modified node
+  */ 
+  transformNode(node) {
+    return node
+  }
+
+  transformNodeValues(nodes) {
+    this.nodes = nodes
+    let newNodes = {}
+    Object.entries(nodes).forEach(([key, node]) => {
+      newNodes[key] = this.transformNode(node)
+    })
+
+    return newNodes
+  }
+
+  /********************************************************************************
+   *
+   *
+   * COLOR
+   * 
+   * 
+  ********************************************************************************/
+  get colorScale() {
+    return 'none'
+  }
+
+  getRawColorValue(node) {
+    return node[this.key]
+  }
+
+  getRawColorValues(nodes) {
+    return Object.entries(nodes).map(node => this.getRawColorValue(node))
+  }
+
+  getScaledColorValue(node, scale) {
+    let raw = this.getRawColorValue(node)
+    let scaled = scale(raw)
+    return d3.hsl(210 * (1 - raw), 0.7, 0.5);
+  }
+
+  getScaledColorValues(nodes, scale) {
+    return nodes.map(node => getScaledColorValue(node, scale))
+  }
+
+  /********************************************************************************
+   *
+   *
+   * SIZE
+   * 
+   * 
+  ********************************************************************************/
[…]
