+++ client/js/babel.config.js
@@ -2,4 +2,8 @@ module.exports = {
   presets: [
     '@babel/preset-env',
   ],
+
+  plugins: [
+    "@babel/plugin-proposal-class-properties"
+  ]
 };

[…]
