+++ client/js/attribute.js
@@ -1,5 +1,8 @@
 import * as d3 from 'd3'
 
+const zip = (arr, ...arrs) => {
+  return arr.map((val, i) => arrs.reduce((a, arr) => [...a, arr[i]], [val]))
+}
 /********************************************************************************
  * Attribute object
  *
@@ -13,16 +16,80 @@ import * as d3 from 'd3'
  *    - slapping them onto the nodes
 ********************************************************************************/
 export class Attribute {
-  static TYPE = 'none'
 
-  static defaultColorValue = d3.rgb(100, 100, 100)
-  static defaultValue = 1
+  get defaultColorValue() { return d3.rgb(128, 128, 128) }
+  get defaultValue() { return 1 }
+  static get DISPLAY_COLOR() { return true }
+  static get DISPLAY_SIZE() { return true }
 
-  static DISPLAY_COLOR = true
-  static DISPLAY_SIZE = true
+  get TYPE() {
+    return 'none'
+  }
 
   get sizeScale() {
     return 'nodeSize'
+  }
+
+  getLegendValuesAndText(rawValues, scaledValues, nodes) {
+    let text = rawValues
+    let values = scaledValues
+    return this.legendSort(values, text)
+  }
+
+  legendSort(rawValues, rawText) {
+    if (rawValues == undefined) {
+      rawValues = []
+    }
+    if (rawText == undefined) {
+      rawText = []
+    }
+
+    let length = rawValues.length < rawText.length ? rawValues.length : rawText.length
+    let pairList = []
+
+    let seenText = {}
+
+    for (let i = 0; i < length; i++) {
+      let text = rawText[i]
+      if (seenText[text] == undefined) {
+        seenText[text] = true
+        pairList.push({
+          'value': rawValues[i],
+          'text': text,
+        })
+      }
+    }
+
+    // sort by text element
+    pairList.sort((a, b) => a.text - b.text)
+
+    return {
+      'values': pairList.map(x => x.value),
+      'text': pairList.map(x => x.text),
+    }
+  }
+
+  // 'bins' a set of values so that we can display a finite legend.
+  makeBinnedLegend(rawValues, bins) {
+    if (bins == undefined || isNaN(bins)) {
+      bins = 4
+    }
+
+    let [min, max] = [d3.min(rawValues), d3.max(rawValues)]
+    let step = (max - min) / bins
+
+    let values = []
+    
+    values.push(rounddown(min, 10))
+    for (let i = 1; i < bins; i++) {
+      values.push(round(min + i * step, 10))
+    }
+    values.push(roundup(max, 10))
+
+    return {
+      'values': values,
+      'text': values.slice(0),
+    }
   }
 
   // want to opaquly handle color and size
@@ -49,11 +116,11 @@ export class Attribute {
   // EXTENT and other size properties
   extent(nodes) {
     let nodeValues = Object.values(nodes).map(node => node[this.key])
-    return d3.set(nodeValues).values().sort()
+    return d3.set(nodeValues).values()
   }
 
   colorExtent(nodes) {
-    return d3.set(this.getRawColorValues(nodes)).values().sort()
+    return d3.set(this.getRawColorValues(nodes)).values()
   }
 
   valueSetSize(nodes) {
@@ -172,37 +239,86 @@ export class Attribute {
 }
 
 export class NameAttribute extends Attribute {
-  static TYPE = 'name'
+  get TYPE() {
+    return 'name'
+  }
 
-  static DISPLAY_COLOR = false
-  static DISPLAY_SIZE = false
+  static get DISPLAY_COLOR() { return false }
+  static get DISPLAY_SIZE() { return false }
 
 }
[…]
