+++ client/js/webweb.bundle.js
@@ -170,6 +170,17 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 
 /***/ }),
 
+/***/ "./listeners.js":
+/*!**********************!*\
+  !*** ./listeners.js ***!
+  \**********************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+eval("throw new Error(\"Module build failed (from ./node_modules/babel-loader/lib/index.js):\\nSyntaxError: /Users/hne/Documents/projects/webweb/client/js/listeners.js: Unexpected token, expected \\\"{\\\" (1:28)\\n\\n\\u001b[0m\\u001b[31m\\u001b[1m>\\u001b[22m\\u001b[39m\\u001b[90m 1 | \\u001b[39m\\u001b[36mexport\\u001b[39m \\u001b[36mclass\\u001b[39m \\u001b[33mGlobalListeners\\u001b[39m() {\\u001b[0m\\n\\u001b[0m \\u001b[90m   | \\u001b[39m                            \\u001b[31m\\u001b[1m^\\u001b[22m\\u001b[39m\\u001b[0m\\n\\u001b[0m \\u001b[90m 2 | \\u001b[39m  constructor(settings\\u001b[33m,\\u001b[39m callHandler) {\\u001b[0m\\n\\u001b[0m \\u001b[90m 3 | \\u001b[39m    \\u001b[36mthis\\u001b[39m\\u001b[33m.\\u001b[39mcallHandler \\u001b[33m=\\u001b[39m callHandler\\u001b[0m\\n\\u001b[0m \\u001b[90m 4 | \\u001b[39m    \\u001b[36mthis\\u001b[39m\\u001b[33m.\\u001b[39msettings \\u001b[33m=\\u001b[39m settings\\u001b[0m\\n    at Parser.raise (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:6325:17)\\n    at Parser.unexpected (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:7642:16)\\n    at Parser.expect (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:7628:28)\\n    at Parser.parseClassBody (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:10599:10)\\n    at Parser.parseClass (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:10574:22)\\n    at Parser.parseStatementContent (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:9871:21)\\n    at Parser.parseStatement (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:9829:17)\\n    at Parser.parseExportDeclaration (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:11017:17)\\n    at Parser.maybeParseExportDeclaration (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:10967:31)\\n    at Parser.parseExport (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:10896:29)\\n    at Parser.parseStatementContent (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:9933:27)\\n    at Parser.parseStatement (/Users/hne/Documents/projects/webweb/client/js/node_modules/@babel/parser/lib/index.js:9829:17)\\n    at P
[…]
