### 2019-08-12 #1458 @hneutr
Hi @cMadan!

- There's a tagged release [here](https://github.com/dblarremore/webweb/releases/tag/0.0.37). The released version is `0.0.37`.
- [Here](https://zenodo.org/record/3366140#.XVF1yZNKjRY) is the zenodo upload! Author information matches.
- The zenodo DOI is: 10.5281/zenodo.3366140

Is there anything else that needs to be done before publication?

Best,
Hunter
