### 2019-08-12 #1458 @cMadan
@hneutr, great! I think you're all good!

@openjournals/joss-eics, we're all set to accept here!
