+++ paper/paper.bib
@@ -28,7 +28,7 @@
     volume = 7,
     address = {Pasadena, CA USA},
     month = 8,
-    organization = {SciPi2008}
+    organization = {SciPy2008}
 }
 
 @article{cytoscape,
