### 2019-08-12 #1458 @danielskatz
there's a bib typo, fixed in https://github.com/dblarremore/webweb/pull/53
Please merge this or otherwise fix it, then we can finalize the acceptance.
